+++ semcore/data-table/__tests__/data-table.browser-test.tsx
@@ -10,4 +10,14 @@ test.describe('DataTable', () => {
 
     await expect(page).toHaveScreenshot();
   });
+
+  test('Renders correctly with scroll if some columns are fixed', async ({ page }) => {
+    const standPath = 'semcore/data-table/__tests__/stands/scroll-and-fixed-column.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    await page.setViewportSize({ width: 820, height: 500 });
+
+    await expect(page).toHaveScreenshot();
+  });
 });