+++ website/docs/components/slider/examples/numeric_slider.tsx
@@ -3,6 +3,7 @@ import Slider from 'intergalactic/slider';
 import InputNumber from 'intergalactic/input-number';
 import Tooltip from 'intergalactic/tooltip';
 import { Box } from 'intergalactic/flex-box';
+import { Text } from 'intergalactic/typography';
 
 const Demo = () => {
   const [value, setValue] = React.useState(51);
@@ -22,24 +23,46 @@ const Demo = () => {
 
   return (
     <>
+      <Text tag='label' size={200} htmlFor='slider-represantation'>
+        Slider representation
+      </Text>
       <Box w={140}>
-        <Slider mb={3} value={value} onChange={setValue} step={1} min={min} max={max}>
+        <Slider
+          id='slider-represantation'
+          aria-label='Number slider representation'
+          mb={3}
+          value={value}
+          onChange={setValue}
+          step={1}
+          min={min}
+          max={max}
+        >
           <Slider.Bar />
           <Slider.Knob />
         </Slider>
       </Box>
-      <Tooltip
-        title={`Please enter a valid value within ${min} and ${max}.`}
-        visible={!!error}
-        interaction='click'
-        theme='warning'
-        placement='right'
-      >
-        <InputNumber mt={4} w={140} size='m' state={error ? 'invalid' : 'normal'}>
-          <InputNumber.Value step={1} value={value.toString()} onChange={handleInput} />
-          <InputNumber.Controls showControls />
-        </InputNumber>
-      </Tooltip>
+      <Text tag='label' size={200} htmlFor='numeric-value-represantation'>
+        Numeric value representation
+      </Text>
+      <Box mt={2}>
+        <Tooltip
+          title={`Please enter a valid value within ${min} and ${max}.`}
+          visible={!!error}
+          interaction='click'
+          theme='warning'
+          placement='right'
+        >
+          <InputNumber mt={2} w={140} size='m' state={error ? 'invalid' : 'normal'}>
+            <InputNumber.Value
+              id='numeric-value-represantation'
+              step={1}
+              value={value.toString()}
+              onChange={handleInput}
+            />
+            <InputNumber.Controls showControls />
+          </InputNumber>
+        </Tooltip>
+      </Box>
     </>
   );
 };