+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.49.0] - 2024-07-13
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/animation` [2.29.0 ~> 2.30.0], `@semcore/utils` [4.30.0 ~> 4.31.0], `@semcore/core` [2.27.0 ~> 2.28.0]).
+
 ## [3.48.1] - 2024-07-05
 
 ### Changed